test(tools): extract shared J-lens toy-bridge fixtures into conftest
- move D_MODEL/N_LAYERS/D_VOCAB/SEQ_LEN/SKIP_FIRST/CORPUS, _ToyBlock,
  _CausalSumBlock, _ToyTokenizer, _ToyBridge, _NotABridge, _lens, and the
  toy_model fixture from test_jacobian_lens.py into conftest.py
- import the shared symbols back into test_jacobian_lens.py and remove the
  now-unused contextmanager and HookPoint imports
- centralize the shared test setup so additional hook tests can reuse the
  same fixtures without duplicating roughly 90 lines of test code
